removed test_activation.py and moved to test_jax_nn_activations.py
